+++ examples/mpi_vegas_example.cpp
@@ -58,8 +58,8 @@ int main(int argc, char* argv[])
 	{
 		// print the cumulative result
 		std::cout << ">> cumulative result (excluding first iteration):\n>> N="
-			<< result.calls << " I=" << result.value << " +- " << result.error
-			<< " chi^2/dof=" << chi_square_dof << "\n";
+			<< result.calls() << " I=" << result.value() << " +- "
+			<< result.error() << " chi^2/dof=" << chi_square_dof << "\n";
 	}
 
 	// clean up