+++ tests/test_mc_helper.cpp
@@ -9,32 +9,6 @@ typedef testing::Types<float, double, long double> MyT;
 template <typename T> class McHelper : public testing::Test { };
 TYPED_TEST_CASE(McHelper, MyT);
 
-TYPED_TEST(McHelper, CumulativeResult0WithZero)
-{
-	typedef TypeParam T;
-
-	std::vector<hep::mc_result<T>> zero_results;
-	auto result = hep::cumulative_result0(zero_results.begin(), zero_results.end());
-
-	// assert that results are reasonable even if there is no real input
-	EXPECT_EQ( 0U  , result.calls );
-	EXPECT_EQ( T() , result.value );
-	EXPECT_EQ( T() , result.error );
-}
-
-TYPED_TEST(McHelper, CumulativeResult1WithZero)
-{
-	typedef TypeParam T;
-
-	std::vector<hep::mc_result<T>> zero_results;
-	auto result = hep::cumulative_result1(zero_results.begin(), zero_results.end());
-
-	// assert that results are reasonable even if there is no real input
-	EXPECT_EQ( 0U  , result.calls );
-	EXPECT_EQ( T() , result.value );
-	EXPECT_EQ( T() , result.error );
-}
-
 TYPED_TEST(McHelper, CumulativeResult0WithOne)
 {
 	typedef TypeParam T;
@@ -44,9 +18,9 @@ TYPED_TEST(McHelper, CumulativeResult0WithOne)
 	};
 	auto result = hep::cumulative_result0(one_result.begin(), one_result.end());
 
-	EXPECT_EQ( one_result.front().calls , result.calls );
-	EXPECT_NEAR( one_result.front().value , result.value , T(1e-10) );
-	EXPECT_NEAR( one_result.front().error , result.error , T(1e-10) );
+	EXPECT_EQ( one_result.front().calls() , result.calls() );
+	EXPECT_NEAR( one_result.front().value() , result.value() , T(1e-10) );
+	EXPECT_NEAR( one_result.front().error() , result.error() , T(1e-10) );
 }
 
 TYPED_TEST(McHelper, CumulativeResult1WithOne)
@@ -58,9 +32,9 @@ TYPED_TEST(McHelper, CumulativeResult1WithOne)
 	};
 	auto result = hep::cumulative_result1(one_result.begin(), one_result.end());
 
-	EXPECT_EQ( one_result.front().calls , result.calls );
-	EXPECT_NEAR( one_result.front().value , result.value , T(1e-10) );
-	EXPECT_NEAR( one_result.front().error , result.error , T(1e-10) );
+	EXPECT_EQ( one_result.front().calls() , result.calls() );
+	EXPECT_NEAR( one_result.front().value() , result.value() , T(1e-10) );
+	EXPECT_NEAR( one_result.front().error() , result.error() , T(1e-10) );
 }
 
 TYPED_TEST(McHelper, CumulativeResult0WithTwo)
@@ -72,14 +46,16 @@ TYPED_TEST(McHelper, CumulativeResult0WithTwo)
 		hep::mc_result<T>(100, T(100.0), T(200.0)),
 		hep::mc_result<T>(100, T(100.0), T(200.0))
 	};
-	auto result = hep::cumulative_result0(two_results.begin(), two_results.end());
+	auto result = hep::cumulative_result0(two_results.begin(),
+		two_results.end());
 
 	// should give two times the calls
-	EXPECT_EQ( 2 * two_results.front().calls , result.calls );
+	EXPECT_EQ( 2 * two_results.front().calls() , result.calls() );
 	// the same result and
-	EXPECT_NEAR( two_results.front().value , result.value , T(1e-10) );
+	EXPECT_NEAR( two_results.front().value() , result.value() , T(1e-10) );
 	// an error reduces by 1/sqrt(2)
-	EXPECT_NEAR( two_results.front().error / std::sqrt(T(2.0)) , result.error , T(1e-9) );
+	EXPECT_NEAR( two_results.front().error() / std::sqrt(T(2.0)) ,
+		result.error() , T(1e-9) );
 }
 
 TYPED_TEST(McHelper, CumulativeResult1WithTwo)
@@ -91,14 +67,15 @@ TYPED_TEST(McHelper, CumulativeResult1WithTwo)
 		hep::mc_result<T>(100, T(100.0), T(200.0)),
 		hep::mc_result<T>(100, T(100.0), T(200.0))
 	};
-	auto result = hep::cumulative_result1(two_results.begin(), two_results.end());
+	auto result = hep::cumulative_result1(two_results.begin(),
+		two_results.end());
 
 	// should give two times the calls
-	EXPECT_EQ( 2 * two_results.front().calls , result.calls );
+	EXPECT_EQ( 2 * two_results.front().calls() , result.calls() );
 	// the same result and
-	EXPECT_NEAR( two_results.front().value , result.value , T(1e-10) );
+	EXPECT_NEAR( two_results.front().value() , result.value() , T(1e-10) );
 	// zero error
-	EXPECT_EQ( T() , result.error );
+	EXPECT_EQ( T() , result.error() );
 }
 
 TYPED_TEST(McHelper, CumulativeResult1WithThree)
@@ -111,14 +88,15 @@ TYPED_TEST(McHelper, CumulativeResult1WithThree)
 		hep::mc_result<T>(100, T(100.0), T(200.0)),
 		hep::mc_result<T>(100, T(100.0), T(200.0)),
 	};
-	auto result = hep::cumulative_result1(two_results.begin(), two_results.end());
+	auto result = hep::cumulative_result1(two_results.begin(),
+		two_results.end());
 
 	// should give three times the calls
-	EXPECT_EQ( 3 * two_results.front().calls , result.calls );
+	EXPECT_EQ( 3 * two_results.front().calls() , result.calls() );
 	// the same result and
-	EXPECT_NEAR( two_results.front().value , result.value , T(1e-10) );
+	EXPECT_NEAR( two_results.front().value() , result.value() , T(1e-10) );
 	// zero error
-	EXPECT_EQ( T() , result.error );
+	EXPECT_EQ( T() , result.error() );
 }
 
 TYPED_TEST(McHelper, ChiSquareDof0WithZero)